refactor: Input컴포넌트 분리 및 Todos 수정 Input 컴포넌트는 한곳에서 쓰니까 todo 폴더로 분리했습니다 파일명을 Todos로 변경했는데 안에 함수네이밍은 그대로였네요 수정했습니다!
